02-app/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.3
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
appdirs==1.4.4
apprise==1.9.0
asgi-lifespan==2.1.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
backoff==2.2.1
black==24.10.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coolname==2.2.0
croniter==3.0.3
cryptography==43.0.1
dataclasses-json==0.6.7
datasets==3.0.1
dateparser==1.2.0
db-dtypes==1.3.0
dill==0.3.8
distro==1.9.0
docker==7.1.0
elastic-transport==8.15.0
elasticsearch==8.15.1
elasticsearch-haystack==1.0.0
exceptiongroup==1.2.2
fastapi==0.115.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.20.0
google-api-python-client==2.148.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.20.2
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
graphviz==0.20.3
greenlet==3.1.1
griffe==1.3.2
grpc-google-iam-v1==0.13.1
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
h2==4.1.0
haystack-ai==2.6.0
haystack-experimental==0.2.0
hpack==4.0.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.1
humanize==4.11.0
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.4
jinja2-humanize-extension==0.4.0
jiter==0.6.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.3.2
langchain-community==0.3.1
langchain-core==0.3.9
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
langsmith==0.1.132
lazy-imports==0.3.1
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.0
marshmallow==3.22.0
mdurl==0.1.2
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.9.1
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
openai==1.51.1
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pandas-gbq==0.23.2
pathspec==0.12.1
pendulum==3.0.0
pillow==10.4.0
platformdirs==4.3.6
posthog==3.7.0
prometheus_client==0.21.0
proto-plus==1.24.0
protobuf==5.28.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydata-google-auth==1.8.2
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
ragas==0.1.16
ragas-haystack==1.0.1
readchar==4.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
streamlit==1.39.0
sympy==1.12
tenacity==8.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.8.0
time-machine==2.15.0
tokenizers==0.20.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
transformers==4.45.2
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.31.0
watchdog==5.0.3
websockets==13.1
xxhash==3.5.0
yarl==1.13.1
